Hardcode close loop file descriptor limit to 1024.
Starting a container with sysvinit and start-stop-daemon
may result in a busy close loop blocking container startup
and maxing out CPU if the maximum open file descriptor limit is unlimited.

Limiting the container file descriptor limit with 'ulimit -a'
is not an option for multi-process containers.
Every init script within a container that uses 'start-stop-daemon'
should set the file descriptor limit.

WARNING This may lead to file-descriptor leaks, and is only meant
as a reminder and temporary fix.
A better idea is to loop over the actual existing file descriptors in
/proc/self/fd.

Use the following implementation
https://github.com/containers/conmon/blob/911c786975b7f7217166cbd38e3f20d594ba0fdc/src/close_fds.c#L72